requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
numpy==1.20.1
pandas==1.2.3
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
schedule==1.0.0
six==1.15.0
urllib3==1.26.4